Financial Agent/requirements.txt
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
groq
agno
google
ddgs